+++ src/Lib/RequestTrait.php
@@ -3,7 +3,6 @@
 namespace Zainpay\SDK\Lib;
 
 use GuzzleHttp\Client;
-use GuzzleHttp\Exception\GuzzleException;
 use UnexpectedValueException;
 use Zainpay\SDK\Engine;
 use Zainpay\SDK\Response;
@@ -76,11 +75,17 @@ trait RequestTrait
             ], $headers)
         ]);
 
-        $response = $client->post($url, [
-            'json' => $data,
-        ]);
+        try {
+            $response = $client->post($url, [
+                'json' => $data,
+            ]);
 
-        return new Response($response);
+            return new Response($response);
+        } catch (\GuzzleHttp\Exception\ServerException $e) {
+            return self::HandleServerExceptionResponse($e);
+        } catch (\GuzzleHttp\Exception\RequestException $e) {
+            return self::HandleClientExceptionResponse($e);
+        }
     }
 
     /**
@@ -96,11 +101,17 @@ trait RequestTrait
             ], $headers)
         ]);
 
-        $response = $client->get($url, [
-            'query' => $params,
-        ]);
+        try {
+            $response = $client->get($url, [
+                'query' => $params,
+            ]);
 
-        return new Response($response);
+            return new Response($response);
+        } catch (\GuzzleHttp\Exception\ServerException $e) {
+            return self::HandleServerExceptionResponse($e);
+        } catch (\GuzzleHttp\Exception\RequestException $e) {
+            return self::HandleClientExceptionResponse($e);
+        }
     }
 
 
@@ -117,16 +128,38 @@ trait RequestTrait
             ], $headers)
         ]);
 
-        $response = $client->patch($url, [
-            'json' => $data,
-        ]);
+        try {
+            $response = $client->patch($url, [
+                'json' => $data,
+            ]);
 
-        return new Response($response);
+            return new Response($response);
+        } catch (\GuzzleHttp\Exception\ServerException $e) {
+            return self::HandleServerExceptionResponse($e);
+        } catch (\GuzzleHttp\Exception\RequestException $e) {
+            return self::HandleClientExceptionResponse($e);
+        }
     }
 
 
     protected function createClient(array $config): Client
     {
         return new Client($config);
+    }
+
+    private static function HandleClientExceptionResponse(\GuzzleHttp\Exception\ClientException $e)
+    {
+        $errorResponse = new Response($e->getResponse() ?? null);
+        $errorResponse->setError(true);
+        $errorResponse->setErrorMessage($e->getResponse()->getReasonPhrase());
+        return $errorResponse;
+    }
+
+    private static function HandleServerExceptionResponse(\GuzzleHttp\Exception\ServerException $e)
+    {
+        $errorResponse = new Response($e->getResponse() ?? null);
+        $errorResponse->setError(true);
+        $errorResponse->setErrorMessage($e->getResponse()->getReasonPhrase());
+        return $errorResponse;
     }
 }