+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [2.0.1] - 2024-01-18
+## Changes
+- fixed payment collected summary by merchant
+- added endpoint for repushing deposit transaction
+- added repush deposit event to the documentation
+
+## [2.0.0] - 2024-01-01
+## Changes
+- added exception handling to allow user to access response object when statusCode is not 200 | OK
+- added error message to allow user get error description when the response has no body
+- updated Documentation
+- updated fund transfer to accomodate callbackUrl
+
 ## [1.0.9] - 2023-11-06
 ## Changes
 - added endpoints for getting card txn history by merchant or zainbox